+++ modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/portlet/action/BaseUADMVCActionCommand.java
@@ -19,10 +19,10 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.portlet.bridges.mvc.BaseMVCActionCommand;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.Portal;
 import com.liferay.user.associated.data.aggregator.UADAggregator;
 import com.liferay.user.associated.data.anonymizer.UADAnonymizer;
 import com.liferay.user.associated.data.web.internal.registry.UADRegistry;
+import com.liferay.user.associated.data.web.internal.util.SelectedUserHelper;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -49,14 +49,14 @@ public abstract class BaseUADMVCActionCommand extends BaseMVCActionCommand {
 	protected List<Object> getEntities(ActionRequest actionRequest)
 		throws Exception {
 
-		UADAggregator uadAggregator = getUADAggregator(actionRequest);
+		List<Object> entities = new ArrayList<>();
 
 		String[] primaryKeys = ParamUtil.getStringValues(
 			actionRequest, "primaryKeys");
 
-		List<Object> entities = new ArrayList<>();
-
 		for (String primaryKey : primaryKeys) {
+			UADAggregator uadAggregator = getUADAggregator(actionRequest);
+
 			entities.add(uadAggregator.get(primaryKey));
 		}
 
@@ -74,15 +74,13 @@ public abstract class BaseUADMVCActionCommand extends BaseMVCActionCommand {
 	protected User getSelectedUser(ActionRequest actionRequest)
 		throws PortalException {
 
-		return portal.getSelectedUser(actionRequest);
+		return selectedUserHelper.getSelectedUser(actionRequest);
 	}
 
 	protected long getSelectedUserId(ActionRequest actionRequest)
 		throws PortalException {
 
-		User selectedUser = portal.getSelectedUser(actionRequest);
-
-		return selectedUser.getUserId();
+		return selectedUserHelper.getSelectedUserId(actionRequest);
 	}
 
 	protected UADAggregator getUADAggregator(ActionRequest actionRequest) {
@@ -98,7 +96,7 @@ public abstract class BaseUADMVCActionCommand extends BaseMVCActionCommand {
 	}
 
 	@Reference
-	protected Portal portal;
+	protected SelectedUserHelper selectedUserHelper;
 
 	@Reference
 	protected UADRegistry uadRegistry;