+++ modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/portlet/action/ViewUADEntitiesMVCRenderCommand.java
@@ -20,6 +20,7 @@ import com.liferay.frontend.taglib.clay.servlet.taglib.util.NavigationItemList;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.dao.search.EmptyOnClickRowChecker;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
@@ -35,6 +36,7 @@ import com.liferay.user.associated.data.web.internal.constants.UADWebKeys;
 import com.liferay.user.associated.data.web.internal.display.UADEntity;
 import com.liferay.user.associated.data.web.internal.display.ViewUADEntitiesDisplay;
 import com.liferay.user.associated.data.web.internal.registry.UADRegistry;
+import com.liferay.user.associated.data.web.internal.util.SelectedUserHelper;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,6 +47,8 @@ import javax.portlet.PortletException;
 import javax.portlet.PortletURL;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -68,7 +72,8 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 		throws PortletException {
 
 		try {
-			User selectedUser = _portal.getSelectedUser(renderRequest);
+			User selectedUser = _selectedUserHelper.getSelectedUser(
+				renderRequest);
 
 			String applicationName = ParamUtil.getString(
 				renderRequest, "applicationName");
@@ -147,9 +152,10 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 	private DropdownItemList _getActionDropdownItems(
 		RenderRequest renderRequest, RenderResponse renderResponse) {
 
-		return new DropdownItemList(
-			_portal.getHttpServletRequest(renderRequest)) {
+		HttpServletRequest request = _portal.getHttpServletRequest(
+			renderRequest);
 
+		return new DropdownItemList() {
 			{
 				add(
 					dropdownItem -> {
@@ -157,7 +163,8 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 							StringBundler.concat(
 								"javascript:", renderResponse.getNamespace(),
 								"doAnonymizeMultiple();"));
-						dropdownItem.setLabel("anonymize");
+						dropdownItem.setLabel(
+							LanguageUtil.get(request, "anonymize"));
 					});
 				add(
 					dropdownItem -> {
@@ -165,7 +172,8 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 							StringBundler.concat(
 								"javascript:", renderResponse.getNamespace(),
 								"doDeleteMultiple();"));
-						dropdownItem.setLabel("delete");
+						dropdownItem.setLabel(
+							LanguageUtil.get(request, "delete"));
 					});
 			}
 
@@ -240,6 +248,9 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 
 	@Reference
 	private Portal _portal;
+
+	@Reference
+	private SelectedUserHelper _selectedUserHelper;
 
 	@Reference
 	private UADRegistry _uadRegistry;