+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/elements/CommandPoshiElement.java
@@ -191,6 +191,14 @@ public class CommandPoshiElement extends PoshiElement {
 				continue;
 			}
 
+			if (isMultilineReadableSyntaxComment(item)) {
+				item = item.replaceFirst("\t", pad + "\t");
+
+				sb.append(item);
+
+				continue;
+			}
+
 			item = item.replaceAll("\n", "\n" + pad);
 
 			sb.append(item.replaceAll("\n\t\n", "\n\n"));