+++ modules/test/poshi-runner/poshi-runner/src/main/java/com/liferay/poshi/runner/elements/PoshiElement.java
@@ -65,6 +65,10 @@ public abstract class PoshiElement
 			return true;
 		}
 
+		if (isMultilineReadableSyntaxComment(readableSyntax)) {
+			return true;
+		}
+
 		return false;
 	}
 
@@ -244,6 +248,16 @@ public abstract class PoshiElement
 
 	protected boolean isElementType(String name, Element element) {
 		if (name.equals(element.getName())) {
+			return true;
+		}
+
+		return false;
+	}
+
+	protected boolean isMultilineReadableSyntaxComment(String readableSyntax) {
+		readableSyntax = readableSyntax.trim();
+
+		if (readableSyntax.endsWith("*/") && readableSyntax.startsWith("/*")) {
 			return true;
 		}